migrate to ai-sdk and refactor chat code
