+++ package.json
@@ -19,6 +19,7 @@
     "with-env": "dotenv --"
   },
   "dependencies": {
+    "@ai-sdk/react": "^1.2.12",
     "@heroicons/react": "^2.1.1",
     "@hookform/resolvers": "^5.1.1",
     "@mdx-js/loader": "^3.0.1",
@@ -36,11 +37,13 @@
     "@tanstack/react-query": "^5.74.4",
     "@types/mdx": "^2.0.13",
     "@upstash/redis": "^1.34.3",
+    "ai": "^4.3.16",
     "better-auth": "^1.2.9",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "embla-carousel-react": "^8.6.0",
+    "fast-deep-equal": "^3.1.3",
     "fuse.js": "^7.0.0",
     "html2canvas": "^1.4.1",
     "lucide-react": "^0.503.0",
