+++ src/styles/globals.css
@@ -147,11 +147,25 @@
 }
 
 @theme inline {
-  --font-sans: var(--font-ibm-plex-sans-arabic), var(--font-plus-jakarta-sans),
+  --font-sans:
+    var(--font-ibm-plex-sans-arabic), var(--font-plus-jakarta-sans),
     ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
     "Segoe UI Symbol", "Noto Color Emoji";
-  --font-scheherazade: var(--font-scheherazade-base), ui-serif, Georgia, Cambria,
+  --font-scheherazade:
+    var(--font-scheherazade-base), ui-serif, Georgia, Cambria,
     "Times New Roman", Times, serif;
+
+  --animate-shiny-text: shiny-text 3s infinite;
+
+  @keyframes shiny-text {
+    0% {
+      background-position: calc(-100% - var(--shiny-width)) 0;
+    }
+
+    100% {
+      background-position: calc(100% + var(--shiny-width)) 0;
+    }
+  }
 }
 
 *,
